adding support for window resizing

+++ L_ETC-core/src/graphics/Vulkan/VulkanSwapChain.cpp
@@ -39,8 +39,11 @@ namespace letc { namespace graphics {
 			return capabilities.currentExtent;
 		}
 		else {
-			//VkExtent2D actualExtent = { WIDTH, HEIGHT };
-			VkExtent2D actualExtent = { width, height};
+
+			VkExtent2D actualExtent = {
+				static_cast<uint32_t>(width),
+				static_cast<uint32_t>(height)
+			};
 
 			actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
 			actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
